Bugfix #5875
The global "total object" counter in Blender kernel was still using short!
Crash can happen when using this counter for allocating selection data.

+++ source/blender/blenkernel/BKE_global.h
@@ -94,8 +94,8 @@ typedef struct Global {
 	struct ListBase recent_files;
     
 	/* totals */
-	short totobj, totlamp, totobjsel, totcurve, totmesh;
-	short totbone, totbonesel;
+	int totobj, totlamp, totobjsel, totcurve, totmesh;
+	int totbone, totbonesel;
 	int totvert, totedge, totface, totvertsel, totedgesel, totfacesel;
     
 	short afbreek, moving;
